Use `page` for news index, `newsitem_page` for paginated page
Jinja2 templates do not support using `self` for a context variable.
Wagtail now uses `page` instead of `self` because of this. `wagtailnews`
now does the same. This conflicted with the `page` variable which used
to be the current page of the paginated news item list. This variable is
now called `newsitem_page`.

This is a backwards incompatible change.